Cast to uint for shift (#58)
Another line with the same problem as in f0dc92dc2dd7c666bd84c23f20c3774c10b46a7b. (see #27)

+++ tile.go
@@ -48,7 +48,7 @@ func (tile *Tile) IsValid() bool {
 	if tile.Zoom > 32 || tile.Zoom < 0 {
 		return false
 	}
-	worldTileSize := int(1) << tile.Zoom
+	worldTileSize := int(1) << uint(tile.Zoom)
 	if tile.X < 0 || tile.X >= worldTileSize ||
 		tile.Y < 0 || tile.Y >= worldTileSize {
 		return false
